changes done

+++ src/main/java/com/example/controller/EmployeeController.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,7 +13,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.example.model.Employee;
 import com.example.services.EmployeeService;
 
@@ -22,27 +23,44 @@ public class EmployeeController {
 	EmployeeService employeeService;
 	
 	@GetMapping("/employees")
-	public List<Employee> getEmployees() {
-		return employeeService.getEmployees();	
+	public ResponseEntity<List<Employee>> getEmployees() {
+		return new ResponseEntity<List<Employee>>(employeeService.getEmployees(),HttpStatus.OK);	
 	}
-	@GetMapping("/employee/{employeeId}")
-	public Optional<Employee> getEmployeeByID(@PathVariable int employeeId) {
-		return employeeService.getEmployeeByID(employeeId);	
+	@GetMapping("/employees/{employeeId}")
+	public ResponseEntity<Employee> getEmployeeByID(@PathVariable int employeeId) {
+		Optional<Employee> employee = employeeService.getEmployeeByID(employeeId);
+		if(employee.isPresent())
+			return new ResponseEntity<Employee>(employee.get(), HttpStatus.OK);
+		else
+			return new ResponseEntity<Employee>(HttpStatus.NOT_FOUND);
 	}
 	
 	
 	@PostMapping("/employees")
-	public Employee createEmployeeDetails(@RequestBody Employee employee) {
-		return employeeService.saveEmployeeDetails(employee);
+	public ResponseEntity<String> createEmployeeDetails(@RequestBody Employee employee) {
+		Employee employeeDetails = employeeService.saveEmployeeDetails(employee);
+		return new ResponseEntity<String>("Employee created with id: " + employeeDetails.getEmployeeId(), HttpStatus.CREATED);
 	}
 	
 	@PutMapping("/employees")
-	public Employee updateOrders(@RequestBody Employee employee) {
-		return employeeService.updateEmployeeDetails(employee);
+	public ResponseEntity<String> updateEmployee(@RequestBody Employee employee) {
+	Optional<Employee> employeeInfo = employeeService.getEmployeeByID(employee.getEmployeeId());
+	if(!employeeInfo.isPresent())
+		return new ResponseEntity<String>(HttpStatus.INTERNAL_SERVER_ERROR);
+	else {
+		Employee employeeDetails = employeeService.updateEmployeeDetails(employee);
+		return new ResponseEntity<String>("Updated Employee with id : "+ employeeDetails.getEmployeeId(), HttpStatus.OK);
+	}
 	}
 	@DeleteMapping("/employees/{employeeId}")
-	public String deleteEmployee(@PathVariable int employeeId) {
-		return employeeService.deleteEmployeeId(employeeId);
+	public ResponseEntity<String> deleteEmployee(@PathVariable int employeeId) {
+		Optional<Employee> employeeInfo = employeeService.getEmployeeByID(employeeId);
+		if(!employeeInfo.isPresent())
+			return new ResponseEntity<String>(HttpStatus.INTERNAL_SERVER_ERROR);
+		else 
+			employeeService.deleteEmployeeId(employeeId);
+			return new ResponseEntity<String>("Deleted Employee with id: "+ employeeId,HttpStatus.OK);
+		
 	}
 
 }
